annotator: extract shared ReprGuard, panic-safe Debug via try_borrow
Move the thread_local REPR_GUARD + ReprGuard RAII token out of listdef.rs
into a new repr_guard module; listdef and dictdef both import it instead of
dictdef reaching into listdef's namespace.

Replace the `.borrow()` calls in ListDef::fmt and DictDef::fmt with
`try_borrow`, eliding to `<[...]>` / `<{...}>` when a cell is already
mutably borrowed, so formatting a node mid-mutation cannot double-panic.
