+++ majit/majit-translate/src/annotator/listdef.rs
@@ -38,6 +38,8 @@ use std::collections::HashSet;
 use std::fmt;
 use std::rc::{Rc, Weak};
 
+use super::repr_guard::ReprGuard;
+
 use super::bookkeeper::{Bookkeeper, PositionKey};
 use super::model::{AnnotatorError, SomeList, SomeValue, UnionError};
 
@@ -518,47 +520,6 @@ pub struct ListDef {
     pub(crate) inner: Rc<ListDefInner>,
 }
 
-thread_local! {
-    /// Recursion guard for `Debug` of the interior-mutable annotation
-    /// nodes ([`ListDef`] / [`super::dictdef::DictDef`]), whose shared
-    /// `listitem` can point back to a `SomeValue::List` / `Dict` that
-    /// owns them — a legal self-referential annotation such as
-    /// `l = []; l.append(l)`. Mirrors the thread-local `reprdict` guard
-    /// in `SomeObject.__repr__` (model.py:68-90): a node already being
-    /// formatted renders as an elision instead of recursing forever.
-    static REPR_GUARD: RefCell<Vec<usize>> = const { RefCell::new(Vec::new()) };
-}
-
-/// RAII token for [`REPR_GUARD`]. [`ReprGuard::enter`] returns `None`
-/// when `id` is already on the stack (a cycle), signalling the caller
-/// to elide rather than recurse; the pushed id is popped on drop.
-pub(crate) struct ReprGuard(usize);
-
-impl ReprGuard {
-    pub(crate) fn enter(id: usize) -> Option<ReprGuard> {
-        REPR_GUARD.with(|g| {
-            let mut g = g.borrow_mut();
-            if g.contains(&id) {
-                None
-            } else {
-                g.push(id);
-                Some(ReprGuard(id))
-            }
-        })
-    }
-}
-
-impl Drop for ReprGuard {
-    fn drop(&mut self) {
-        REPR_GUARD.with(|g| {
-            let mut g = g.borrow_mut();
-            if let Some(pos) = g.iter().rposition(|&x| x == self.0) {
-                g.remove(pos);
-            }
-        });
-    }
-}
-
 impl fmt::Debug for ListDef {
     /// Parity with `ListDef.__repr__` (listdef.py:175):
     /// `'<[%r]%s%s%s%s>'`, recursion-guarded so a self-referential
@@ -568,8 +529,12 @@ impl fmt::Debug for ListDef {
         let Some(_guard) = ReprGuard::enter(id) else {
             return f.write_str("<[...]>");
         };
-        let li = self.inner.listitem.borrow();
-        let item = li.borrow();
+        let Ok(li) = self.inner.listitem.try_borrow() else {
+            return f.write_str("<[...]>");
+        };
+        let Ok(item) = li.try_borrow() else {
+            return f.write_str("<[...]>");
+        };
         write!(
             f,
             "<[{:?}]{}{}{}{}>",
